Split the integration tests to force them to be run as different execuables when running cargo test. This mitigates the issue where RPM context configuration is incompatible between tests
